Lock UART access to prevent garbled output
Our 433MHz receiving happens in an interrupt, so it's possible to
receive a new Digoo update while printing out the 1-Wire temperature
details. If we always try to print that will result in garbled output,
so lock access to the UART and don't print if we can't get the lock.
Will result in sometimes dropping the received 433 result, but that's
ok.